+++ test/Header/ContentSecurityPolicyTest.php
@@ -187,4 +187,94 @@ class ContentSecurityPolicyTest extends TestCase
             $headers->toString()
         );
     }
+
+    public static function validDirectives()
+    {
+        return [
+            ['child-src', ["'self'"],"Content-Security-Policy: child-src 'self';"],
+            ['manifest-src', ["'self'"], "Content-Security-Policy: manifest-src 'self';"],
+            ['worker-src', ["'self'"], "Content-Security-Policy: worker-src 'self';"],
+            ['prefetch-src', ["'self'"], "Content-Security-Policy: prefetch-src 'self';"],
+            ['script-src-elem', ["'self'"], "Content-Security-Policy: script-src-elem 'self';"],
+            ['script-src-attr', ["'self'"], "Content-Security-Policy: script-src-attr 'self';"],
+            ['style-src-elem', ["'self'"], "Content-Security-Policy: style-src-elem 'self';"],
+            ['style-src-attr', ["'self'"], "Content-Security-Policy: style-src-attr 'self';"],
+            ['base-uri', ["'self'", "'unsafe-inline'"], "Content-Security-Policy: base-uri 'self' 'unsafe-inline';"],
+            ['plugin-types', ['text/csv'], 'Content-Security-Policy: plugin-types text/csv;'],
+            [
+                'form-action',
+                ['http://*.example.com', "'self'"],
+                "Content-Security-Policy: form-action http://*.example.com 'self';"
+            ],
+            [
+                'frame-ancestors',
+                ['http://*.example.com', "'self'"],
+                "Content-Security-Policy: frame-ancestors http://*.example.com 'self';"
+            ],
+            ['navigate-to', ['example.com'], 'Content-Security-Policy: navigate-to example.com;'],
+            ['sandbox', ['allow-forms'], 'Content-Security-Policy: sandbox allow-forms;'],
+
+            // Other directives
+            ['block-all-mixed-content', [], 'Content-Security-Policy: block-all-mixed-content;'],
+            ['require-sri-for', ['script', 'style'], 'Content-Security-Policy: require-sri-for script style;'],
+            ['trusted-types', ['*'], 'Content-Security-Policy: trusted-types *;'],
+            ['upgrade-insecure-requests', [], 'Content-Security-Policy: upgrade-insecure-requests;'],
+        ];
+    }
+
+    /**
+     * @dataProvider validDirectives
+     *
+     * @param string $directive
+     * @param string[] $values
+     * @param string $expected
+     */
+    public function testContentSecurityPolicySetDirectiveThrowsExceptionIfMissingDirectiveNameGiven(
+        $directive,
+        array $values,
+        $expected
+    ) {
+        $csp = new ContentSecurityPolicy();
+        $csp->setDirective($directive, $values);
+
+        self::assertSame($expected, $csp->toString());
+    }
+
+    /**
+     * @dataProvider validDirectives
+     *
+     * @param string $directive
+     * @param string[] $values
+     * @param string $header
+     */
+    public function testFromString($directive, array $values, $header)
+    {
+        $contentSecurityPolicy = ContentSecurityPolicy::fromString($header);
+
+        self::assertArrayHasKey($directive, $contentSecurityPolicy->getDirectives());
+        self::assertSame(implode(' ', $values), $contentSecurityPolicy->getDirectives()[$directive]);
+    }
+
+    /**
+     * @return string
+     */
+    public function directivesWithoutValue()
+    {
+        yield ['block-all-mixed-content'];
+        yield ['upgrade-insecure-requests'];
+    }
+
+    /**
+     * @dataProvider directivesWithoutValue
+     *
+     * @param string $directive
+     */
+    public function testExceptionWhenProvideValueWithDirectiveWithoutValue($directive)
+    {
+        $csp = new ContentSecurityPolicy();
+
+        $this->expectException(InvalidArgumentException::class);
+        $this->expectExceptionMessage($directive);
+        $csp->setDirective($directive, ['something']);
+    }
 }